src/inference/vlm_agent.py
from transformers import AutoModelForCausalLM, AutoTokenizer
import torch
from PIL import Image

class VLMAgent:
    """
    Agent that uses a Vision-Language Model to make traffic control decisions.
    """
    def __init__(self, model_path, device="cuda"):
        self.model_path = model_path
        self.device = device
        self.model = None
        self.tokenizer = None
        
    def load_model(self):
        """Loads the VLM model and tokenizer."""
        print(f"Loading model from {self.model_path}...")
        try:
            # Example loading logic for a generic VLM (like Qwen-VL or LLaVA)
            self.tokenizer = AutoTokenizer.from_pretrained(self.model_path, trust_remote_code=True)
            self.model = AutoModelForCausalLM.from_pretrained(self.model_path, device_map=self.device, trust_remote_code=True).eval()
            print("Model loaded successfully.")
        except Exception as e:
            print(f"Failed to load model: {e}")

    def get_decision(self, image_path: str, prompt: str):
        """
        Generates a decision based on the BEV image and text prompt.
        
        Args:
            image_path: Path to the BEV image.
            prompt: Text prompt describing the task.
            
        Returns:
            The model's text output (the decision).
        """
        if self.model is None:
            raise RuntimeError("Model is not loaded. Call load_model() first.")
            
        # Placeholder for inference logic
        # input_ids = self.tokenizer(prompt, return_tensors='pt').to(self.device)
        # image = Image.open(image_path)
        # ... process image and prompt ...
        # generated_ids = self.model.generate(...)
        # return self.tokenizer.decode(generated_ids)
        
        return "DECISION_PLACEHOLDER"
