requirements.txt
torch
transformers
modelscope
sumolib
traci
numpy
pandas
pillow
scikit-learn
accelerate
sft
# Add other specific dependencies as needed
